+++ common/src/main/java/me/rufia/fightorflight/entity/ai/sensors/FOFSensors.java
@@ -15,7 +15,9 @@ public class FOFSensors {
     public static final SensorType<Sensor<PokemonEntity>> POKEMON_WILD_PROACTIVE = register("pokemon_wild_proactive_fof", PokemonWildProactiveSensor::new);
     public static final SensorType<Sensor<PokemonEntity>> POKEMON_CAUGHT_BY = register("pokemon_caught_by_fof", PokemonCaughtBySensor::new);
     public static final SensorType<Sensor<PokemonEntity>> POKESTAFF_ATTACK_TARGET = register("pokestaff_attack_target_fof", PokeStaffAttackTargetSensor::new);
-    public static final SensorType<Sensor<PokemonEntity>> POKESTAFF_WALK_TARGET=register("pokestaff_walk_target_fof", PokeStaffWalkTargetSensor::new);
+    public static final SensorType<Sensor<PokemonEntity>> POKESTAFF_WALK_TARGET = register("pokestaff_walk_target_fof", PokeStaffWalkTargetSensor::new);
+    public static final SensorType<Sensor<PokemonEntity>> POKESTAFF_AUTO_BATTLE = register("pokestaff_auto_battle_fof", PokeStaffAutoBattleSensor::new);
+
     public static <E extends LivingEntity, U extends Sensor<E>> SensorType<U> register(String id, Supplier<U> supplier) {
         var sensorType = new SensorType<>(supplier);
         sensors.putIfAbsent(id, sensorType);